infra/shared/pyproject.toml
[project]
name = "richmond-shared"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pulumi>=3.0",
    "pulumi-cloudflare>=5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.setuptools]
packages = ["shared"]
package-dir = {"shared" = "."}

[tool.uv.workspace]
members = []
